+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # CHANGELOG.md
 
-# v0.0.1
+## v0.1.0
+
+* Supports the HTTP POST SearchByASICv201408 method including test cases
+* Added support for ACN validation with tests
+
+## v0.0.1
 
 * Initial release
 * Supports the HTTP POST SearchByABNv201408 method
+* Support for ABN validation